firmware: deprecated old control interface
