+++ firmware/smolusb/src/lib.rs
@@ -6,7 +6,7 @@
 
 pub mod class;
 pub mod control;
-pub mod control_new;
+pub mod control_deprecated;
 pub mod descriptor;
 pub mod device;
 pub mod error;
